created two different controllers
PdfController — this controller will have the endpoints related to
converting to Pdf.

WordController — This controller will have the endpoints related to
converting to Word.
